libobs: Rename "source_valid" and display warning
To prevent confusion with the new obs_source_valid function which
displays a warning, rename the "source_valid" function to "data_valid"
to emphasize that it's checking for the validity of the internal data as
well as the source itself.

Now also displays a warning if the source parameter is null.